New variable $config->http_host in basic config
The fixes a problem when using a CompleteSearch instance via a proxy.
The file helper.php uses $_SERVER['HTTP_HOST'] to determine the first
part of the URL. But when using a proxy, this returns the name and port
of the internal server which leads to requests from the frontend being
blocked due to the cross-origin policy.