Test scrollable table, working example
This cost me some nerves. It's surprisingly hard to do this. In
particular, do to it cleanly and elegantly. I think I got it now. I also
documented it extensively, so that it's clear why what needs to be done.

There is also JavaScript code that adds additional rows (more
specifically, doubling their number) when one has scrolled to the
bottom. That, too, was surprisingly non-trivial.