Refactor code with classes HitBox and FacetBox
In the process, fixed a number of small bug of the previous version.
This was a lot of work (more than expected). But now it should be easy
to add more functionality. For example, loading more completions when
the scrollbar in one of the facet boxes hits the bottom.

In the previous version, there were all kinds of global variables and
clases for creating the initial facet boxes, etc. Now each box (hit box
as well as the facet boxes) knows its status, what is displayed in it,
etc. This makes much more sense.