New feature: allow a secondary vocabulary
1. Words from the secondary vocabulary are also allowed as candidates
for partial word matches, but we only find partial word matches in words
from the primary vocabulary. The typical use is to use a general
dictionary for the secondary vocabulary. See PartialWords.cpp for a
detailed explanation of how it works.

2. Improve / extend the documentation in various places.

3. Add variables for a few options (verbosity, output format, min word
length). The verbosity now has three levels. It's still only variables,
we should make these proper command line options.