Fix stop rule in applications/Makefile
The start rule in applications/Makefile uses a custom path for the pid
file. This has the effect that the pid is in the data directory. The
default is in the users home directory.
The stop rule didn't use this custom path and thus fell back to the
default path. This meant that the server couldn't be stopped correctly.