Fix `stop` target in `applications/Makefile` (#5)
The start rule in applications/Makefile uses a custom path for the PID file. This has the effect that the PID file resides in the data directory. By default, the file resides in the user's home directory. The stop rule didn't use this custom path and thus fell back to the default path. This meant that the server couldn't be stopped correctly.